+++ src/navbar/Navbar.jsx
@@ -1,71 +1,46 @@
 import React, { Component } from 'react';
 import { Link } from 'react-router-dom';
 import axios from 'axios';
-import { setAuthToken, unsetAuthToken } from '../login/Login';
 import history from '../history.js';
+import { LoginContext } from '../App';
 
 import './Navbar.css';
 
 class Navbar extends Component {
-  constructor(props) {
-    super(props);
-    this.state = { isAuthed: null };
-  }
-
-  componentDidMount() {
-    setAuthToken();
-    axios
-      .get('/user')
-      .then(() => {
-        this.setState({ isAuthed: true });
-      })
-      .catch(() => {
-        this.setState({ isAuthed: false });
-      });
-  }
-
-  profile() {
-    return (
-      <li>
-        <Link to="/profile">Profile</Link>
-      </li>
-    );
-  }
-
-  addBook() {
-    return (
-      <li>
-        <Link to="/" className="light-btn">
-          Add a Book
-        </Link>
-      </li>
-    );
-  }
-
-  logout() {
-    return (
-      <li>
-        <Link
-          to="/login"
-          className="logout-btn"
-          onClick={() => {
-            unsetAuthToken();
-            axios.get('/logout');
-            history.push('/login');
-          }}
-        >
-          Logout
-        </Link>
-      </li>
-    );
-  }
-
-  login() {
-    return (
-      <li>
-        <Link to="/login">Login</Link>
-      </li>
-    );
+  navItems({ loggedIn, logout }) {
+    if (loggedIn) {
+      return (
+        <React.Fragment>
+          <li>
+            <Link
+              to="/login"
+              className="logout-btn"
+              onClick={() => {
+                logout();
+                axios.get('/logout');
+                history.push('/login');
+              }}
+            >
+              Logout
+            </Link>
+          </li>
+          <li>
+            <Link to="/profile">Profile</Link>
+          </li>
+          <li>
+            <Link to="/" className="light-btn">
+              Add a Book
+            </Link>
+          </li>
+        </React.Fragment>
+      );
+    } else {
+      return (
+        <li>
+          <Link to="/login">Login</Link>
+        </li>
+      );
+    }
   }
 
   render() {
@@ -76,9 +51,7 @@ class Navbar extends Component {
             Asterisk
           </Link>
 
-          {this.state.isAuthed ? this.logout() : this.login()}
-          {this.state.isAuthed ? this.profile() : ''}
-          {this.state.isAuthed ? this.addBook() : ''}
+          <LoginContext.Consumer>{this.navItems}</LoginContext.Consumer>
         </ul>
       </div>
     );
